Modularization of Iter, MethodIter and RangeIter
- Iter, RefValueIter and BoxedRefValueIter remain in values::iter
- MethodIter moved into values::methoditer
- RangerIter moved into builtin::range
