+++ src/value/mod.rs
@@ -3,6 +3,7 @@ pub mod dict;
 pub mod iter;
 pub mod list;
 mod method;
+mod methoditer;
 mod object;
 mod parselet;
 mod refvalue;
@@ -12,9 +13,10 @@ pub mod value;
 
 pub use self::str::Str;
 pub use dict::Dict;
-pub use iter::{Iter, MethodIter};
+pub use iter::{Iter, RefValueIter};
 pub use list::List;
 pub use method::Method;
+pub use methoditer::MethodIter;
 pub use object::{BoxedObject, Object};
 pub(crate) use parselet::{Parselet, ParseletRef};
 pub use refvalue::RefValue;
